@@ -1,11 +1,23 @@
-const isEmpty = require('lodash/isEmpty');
-const filterObject = require('./utils/filter-object');
-const objectMerge = require('./utils/object-merge');
-const objectDifference = require('./utils/object-difference');
-const fillShape = require('./utils/fill-shape');
+import isEmpty from "lodash/isEmpty";
+import { AnyAction, DeepPartial, Reducer, Store, StoreCreator } from 'redux';
+import { defaultParams } from "./main-enhancer";
+import { fillShape } from './utils/fill-shape';
+import { filterObject } from "./utils/filter-object";
+import { objectDifference } from "./utils/object-difference";
+import { objectMerge } from "./utils/object-merge";
+import { Options } from "./types";
+
+
+type Listener = () => void;
+interface Flags {
+  isUpdating: boolean;
+  isDispatching: boolean;
+  forwardOnUpdate: boolean;
+  senderClientId?: string | null;
+}
 
 function getSubscribeFuncs() {
-  let currentListeners = [];
+  let currentListeners: Listener[] = [];
   let nextListeners = currentListeners;
   function ensureCanMutateNextListeners() {
     if (nextListeners === currentListeners) {
@@ -14,7 +26,7 @@ function getSubscribeFuncs() {
   }
 
   return {
-    subscribe: (flags, listener) => {
+    subscribe: (flags: Flags, listener: Listener) => {
       if (typeof listener !== 'function') {
         throw new Error('Expected listener to be a function.');
       }
@@ -46,8 +58,17 @@ function getSubscribeFuncs() {
   };
 }
 
-module.exports = ({ params, flags, storeCreator, reducer, initialState, forwarder }) => {
-  const parsedReducer = (state, action) => {
+export interface SetupElectronStoreContext {
+  params: Options & typeof defaultParams;
+  flags: Flags;
+  storeCreator: StoreCreator;
+  reducer: Reducer<any, any>;
+  initialState?: DeepPartial<any>;
+  forwarder(action: AnyAction): void;
+}
+
+export const setupElectronStore = ({ params, flags, storeCreator, reducer, initialState, forwarder }: SetupElectronStoreContext): Store<any, any> => {
+  const parsedReducer = (state: any, action: AnyAction) => {
     if (flags.isUpdating) {
       flags.isUpdating = false;
       const { updated, deleted } = action.payload;
@@ -61,10 +82,10 @@ module.exports = ({ params, flags, storeCreator, reducer, initialState, forwarde
 
   const store = storeCreator(parsedReducer, initialState);
   const { subscribe, callListeners } = getSubscribeFuncs();
-  store.subscribe = listener => subscribe(flags, listener);
+  store.subscribe = (listener) => subscribe(flags, listener);
 
   const storeDotDispatch = store.dispatch;
-  const doDispatch = (action) => {
+  const doDispatch = (action: AnyAction) => {
     flags.isDispatching = true;
     try {
       params.preDispatchCallback(action);
