@@ -1,13 +1,12 @@
-const isObject = require('lodash/isObject');
-const keys = require('lodash/keys');
+import keys from 'lodash/keys';
+import { isShallow } from './object-difference';
 
-const isShallow = (val) => Array.isArray(val) || !isObject(val);
 
-module.exports = function objectMerge(a, b) {
+export function objectMerge(a: any, b: any) {
   if (a === b || isShallow(a) || isShallow(b)) {
     return b !== undefined ? b : a;
   }
-  let merged = {};
+  let merged: any = {};
   keys(a).forEach(key => merged[key] = objectMerge(a[key], b[key]));
   keys(b).forEach(key => a[key] === undefined && (merged[key] = b[key]));
   return merged;
